Complete playRound, Add playGame function
If playRound is called, it declares a variable "result"
that will store the result of a round, according to
the "rock, paper, scissors" rule. It will log the
result along with the brief process as well.

When playGame is called, it declares humanScore and
computerScore, initiating them with 0. It calls playRound
5 times. Before every invoke, it re-assign player's
choice and computer's choice so that each round results
could differ. It takes the return value of the invoke,
put in into a switch statement, and increment the player
or computer score based on its result.
After 5 times of round, it logs the final result,
summing up every round.